Add 'GET /organizations/:orgName/repositories' to get the list of repositories by organization
This works similarly to the 'GET /repositories' endpoint, except that it
scopes the repos to the specified organization.
The goal is to alleviate the issue we are currently facing in the UI
when there are a lot of repositories, and only partial repositories
would be returned (and the UI deduced the organization list from the
repository list).
Instead, now, there are separate (but still paginated) endpoints to list
the Organizations accessible,
and to get the list of repositories within each organization.

Please note that the organization specified needs to be accessible from
any the configured GitHub Integrations.