Do not create PR if the default branch already contains a catalog-info file
PR was initially being created (and only closed afterwards if it had 0
diff), as we assumed that the bulk import PR could update the existing
catalog-info YAML file.
But with hindsight, the goal of a Bulk Import is just to help onboard
projects. So it should be fine to just import if there is already an
existing catalog-info file in the default branch.
Otherwise, it might be confusing because if we go to the Catalog,
we would see a Component registered in Backstage,
but from the Bulk Import standpoint, it would be waiting for approval.

Co-authored-by: Yi Cai <yicai@redhat.com>