Get catalog-info file name and PR branch from config
This uses the same branch as the one created by the "Register Existing
Component" workflow from Backstage.
This way, if a user registers their component using this component and a
PR is created by Backstage, the repo will still be marked as
'WAIT_PR_APPROVAL' from the bulk import perspective.